style refactor, breaking things into smaller files
