+++ app/.eslintignore
@@ -1,2 +1,3 @@
-app/src/App.js
-app/src/index.js
+node_modules
+src/App.js
+src/index.js
